consultation/requirements.txt
certifi~=2021.5.30
chardet~=4.0.0
click~=8.0.1
colorama~=0.4.4
confluent-kafka~=1.7.0
Flask~=2.0.1
Flask-SQLAlchemy~=2.5.1
greenlet~=1.1.0
idna~=2.10
itsdangerous~=2.0.1
Jinja2~=3.0.1
MarkupSafe~=2.0.1
PyMySQL~=1.0.2
python-dateutil~=2.8.1
requests~=2.25.1
six~=1.16.0
SQLAlchemy~=1.4.20
urllib3~=1.26.6
Werkzeug~=2.0.1
uWSGI~=2.0.19.1
pytest~=6.2.4